Simplify CA renewal
We don't need a rolling update to test the new cert before the rolling upgrade for the new client certs, because the old and new client certs are signed by the same key.

We don't need a rolling update to remove old CA certs from truststores if truststores only ever contain a single CA cert.